feat(v4): add a withTransition mixin and make both transitions generic
Restores v3's structure, where the behaviour lived in a mixin and the
Transition component was withTransition(Base). The earlier port
collapsed the two on the finding that the decorator used `this` for
nothing but reading two options — true at one consumer, false at five:
MenuList, AbstractFigure, FigureVideo and AnchorNavLink each
reimplemented the same state/target/enter/leave/toggle block.

This is v4's first non-service mixin. It reuses createServiceMixin's
type shape (MixedClass) so a consumer threads its own props exactly as
it does for withResize, and its concrete-parameter-plus-cast split for
the same reason core needs it: a class extending a type parameter must
otherwise declare constructor(...args: any[]).

It declares no config. BaseConfig requires a name for the static side
to stay assignable to Base's, and that name would be inherited by any
consumer which forgot its own, registering a component under a name it
never chose — so consumers keep spreading TRANSITION_OPTIONS, the one
line they already had. Core's service mixins declare no config either.

transitionOptions is a new hook, defaulting to $options: it is what
lets MenuList force enterKeep/leaveKeep without reimplementing
enter()/leave(). v3 forced them by overriding the $options getter,
which v4 refuses — $options is a read-only view with no override
point — so the override moves onto the declaration the mixin reads.

ViewTransition gains the same type parameter but stays a direct
Transitionable implementation: the browser owns its animation and its
event names are its own.
