+++ packages/v4/migration/Transition/Transition.ts
@@ -1,63 +1,20 @@
-import { Base } from '../../src/index.js';
-import {
-  enterTransition,
-  leaveTransition,
-  TRANSITION_OPTIONS,
-  type TransitionOptions,
-} from '../../src/utils/transition.js';
+import { Base, type BaseConfig, type BaseProps } from '../../src/index.js';
+import { TRANSITION_OPTIONS } from '../../src/utils/transition.js';
+import { withTransition } from './withTransition.js';
 
-export interface TransitionProps {
-  $options: TransitionOptions;
-  $emits: {
-    'transition-enter': void;
-    'transition-enter-start': void;
-    'transition-enter-end': void;
-    'transition-leave': void;
-    'transition-leave-start': void;
-    'transition-leave-end': void;
-  };
-}
+export type { Transitionable, TransitionInterface, TransitionProps } from './withTransition.js';
 
 /**
- * A component that can be entered and left — what `Dialog` fans its
- * open/close out to.
+ * Runs configured enter and leave CSS transitions on its element.
+ *
+ * Generic and mixin-built, as in v3: the behaviour lives in
+ * `withTransition()` so a component with its own props can mix it in, and
+ * this class is the declarative form of the same thing — `withTransition(Base)`
+ * and nothing else.
  */
-export interface Transitionable {
-  enter(): Promise<void>;
-  leave(): Promise<void>;
-  toggle(): Promise<void>;
-}
-
-/** Runs configured enter and leave CSS transitions on its element. */
-export class Transition extends Base<TransitionProps> implements Transitionable {
-  static config = {
+export class Transition<T extends BaseProps = BaseProps> extends withTransition(Base)<T> {
+  static config: BaseConfig = {
     name: 'Transition',
     options: { ...TRANSITION_OPTIONS },
   };
-
-  state: 'entering' | 'leaving' | null = null;
-
-  get target(): HTMLElement {
-    return this.$el;
-  }
-
-  async enter(): Promise<void> {
-    this.state = 'entering';
-    this.$emit('transition-enter');
-    this.$emit('transition-enter-start');
-    await enterTransition(this.target, this.$options);
-    this.$emit('transition-enter-end');
-  }
-
-  async leave(): Promise<void> {
-    this.state = 'leaving';
-    this.$emit('transition-leave');
-    this.$emit('transition-leave-start');
-    await leaveTransition(this.target, this.$options);
-    this.$emit('transition-leave-end');
-  }
-
-  toggle(): Promise<void> {
-    return this.state === 'entering' ? this.leave() : this.enter();
-  }
 }
