+++ packages/v4/migration/Transition/index.ts
@@ -1,2 +1,8 @@
-export { Transition, type Transitionable, type TransitionProps } from './Transition.js';
+export { Transition } from './Transition.js';
 export { ViewTransition, type ViewTransitionProps } from './ViewTransition.js';
+export {
+  withTransition,
+  type Transitionable,
+  type TransitionInterface,
+  type TransitionProps,
+} from './withTransition.js';
